tasks.tornado tasks
This adds some new tasks:

* TornadoIOLoopTask - runs tornado's IOLoop (in a separate thread)
* TornadoHTTPTask - inits an HTTP Application
* TornadoTask - is a base class for other tasks that require the Tornado
IOLoop.  it provides self.ioloop as well as enforce IOLoopTask is
started

Since only one IOLoop is necessary, TornadoHTTPTask is LOOPLESS

Also, if TornadoIOLoopTask is in the selected services, but no services
actually require it, it will throw SkipTask (and not run).  If you need
the IOLOOP anyway, simply set REQUIRE_TORNADO = True on your VService
subclass