FB303 / Thrift refactor
So, this makes it a little bit harder to create thrift services, but
it's necessary to more cleanly expose fb303 over dbus, and avoid
polluting VServices method signatures with fb303 specific jonx.

So, I've rewritten VFB303Service (the mixin for VService) as a separate
Task.  Maybe this is a bad idea?  If it was, I'll have to make some helper
functions and whatnot to make this jonx more concise later.

A couple of little things have changed and/or improved to make this work.
Most notably:

* FB303::getStatus and FB303::getStatusDetails are now implemented to
return STOPPING and WARNING in some cases.
* VService::getTask() can now take Task Types instead of just the
Tasks' names as strings
* VService now has start_time attribute
* VService::reinitialize() renamed to ::restart()