UniqueQueue: add an `explicit_unsee` attribute.
In addition to the `silent` mode, add an `explicit_unsee` mode to UniqueQueue.

By default, `explicit_unsee` is False, and tems are unseen as soon as
they are popped from the queue.

The problem is that for certain applications, we want to keep the item
as seen until `task_done()` is called.  Since task_done doesn't take an
argument, we add a new function, `unsee(item)` to facilitate this
behavior.

Unittests updated to verify behavior.