QueuTask: prevent unhandled exception from tearing the main loop down
Prior to this patch - raising an exception inside an execute() of a
queued
task when using futures instead of deferreds, would tear down the whole
service thread. This is almost certainly not what we want.

The cleanest way to fix this seems to be in the ExecutionContext, where
we consider an exception handled if it's set on a future, and let the
submitter worry about it.